genius_booker/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
dj-database-url
psycopg2